add synchronization mechanism to MQTT service and clarify MQTT event handler doc
It's proven that MQTT event handler is called from MQTT task. So there must be
synchronization mechanism to protect read and write shared resource in MQTT
service. This commit add this mechanism using mutex in every public API and
MQTT event handler.

This commit also clarifies MQTT event handler that it's not a internal API (
because, MQTT task in ESP-IDF called it)