fix condition to find proper handler for AT command
Last time the condition is to tokenize the AT command family part with delim
'=", which is not right since read command don't use '=' symbol. Also, the test
command does not have any special symbols so this commit changes how to get the
length of AT command family part to compare.