Add parsing test cases for quote and comma in string with some change in PARSETEST cmd
Add new test cases that test intensively the scenario that quoted string
contains quote and comma characters.

As a result, the PARSETEST has one more additional optional quoted string field
to test the case that one optional quoted string in the middle and one optional
quoted string in the end.