AP auto scanner nows only send a pre-configured max number of records
During stress test in 30+ APs nearby, AP auto scanner send a lot of events to
host. Some APs' signal are too weak to even bother, so it's kinda bandwidth-
consuming to send a lot of scanned WiFi APs update. To do it, a new
configuration is introduced to limit total of records that will be sent to host.